Refactor Ditto code
