+++ ditto/src/main/java/eu/arrowhead/core/ditto/service/DittoHttpClient.java
@@ -19,12 +19,12 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 import org.springframework.web.client.RestTemplate;
 import eu.arrowhead.core.ditto.Constants;
 
-@Component
+@Service
 public class DittoHttpClient {
 
 	//=================================================================================================
